Analytics: Metadata SDK improvements; Stub Metadata Plugin related improvements
executeAction() moved from Manager to Plugin.
Manifest for settings changed to declare a "settings" JSON object of type "Group", containing
params and groups.
settingsChanged() notification virtual method added.
Guids changed to string type in Stub Metadata Plugin.

<==== Mercurial data ====>
Branch: vms

+++ common/src/plugins/plugin_manager.cpp
@@ -195,7 +195,7 @@ bool PluginManager::loadNxPlugin(
     {
         // Report settings to the plugin.
         if (!settingsForPlugin.empty())
-            pluginObj->setSettings(&settingsForPlugin[0], settingsForPlugin.size());
+            pluginObj->setSettings(&settingsForPlugin[0], (int) settingsForPlugin.size());
     }
 
     if (auto plugin2Obj = nxpt::ScopedRef<nxpl::Plugin2>(obj->queryInterface(nxpl::IID_Plugin2)))
